# versioning as configurable option