--- a/lib/accept-version.js
+++ b/lib/accept-version.js
@@ -1,0 +1,10 @@
+'use strict'
+
+const SemVerStore = require('semver-store')
+
+module.exports = {
+  storage: SemVerStore,
+  deriveVersion: function (req, ctx) {
+    return req.headers['accept-version']
+  }
+}