hi @delvedor,

This is my current suggestion regarding the problem which I've described here:
https://github.com/delvedor/find-my-way/issues/106

This makes it possible to configure `versions` storage and logic of deriving the version from request configurable by the end user of `find-my-way` (and I hope `fastify` in the future).

This does not affect the default behaviour and benchmarks: https://pastebin.com/hCpa5vzf.

On the negative note: people can write their own "buggy" versioning strategies and blame `find-my-way` author(s) that routing doesn't work ;) 

This PR is still `WIP` since I did not update the documentation and put some simple tests for the custom versioning.

What do you think?

Another question:
How about dropping `node 4`, so [this](https://github.com/delvedor/find-my-way/blob/master/node.js#L17):
```js
function Node (prefix, children, kind, handlers, regex, versions) {}
```

can become this:
```js
function Node ({prefix, children, kind, handlers, regex, versions}) {}
```