@PavelPolyakov 2019-01-02
Good morning and happy NY, @delvedor,

I'd appreciate if you have any feedback on my suggestions.