@PavelPolyakov 2019-01-02
@mcollina thanks for the review.
benchmark results were present in the first comment, here they are inline:

before:
```
> find-my-way@1.17.0 bench /Users/pp/Projects/PavelPolyakov/find-my-way
> node bench.js
 
lookup static route x 29,956,761 ops/sec ±0.99% (90 runs sampled)
lookup dynamic route x 2,301,289 ops/sec ±0.86% (92 runs sampled)
lookup dynamic multi-parametric route x 1,177,672 ops/sec ±1.32% (88 runs sampled)
lookup dynamic multi-parametric route with regex x 1,103,966 ops/sec ±0.86% (91 runs sampled)
lookup long static route x 2,056,904 ops/sec ±0.70% (93 runs sampled)
lookup long dynamic route x 1,423,454 ops/sec ±1.47% (90 runs sampled)
lookup static versioned route x 5,485,497 ops/sec ±1.61% (85 runs sampled)
find static route x 31,022,989 ops/sec ±1.98% (85 runs sampled)
find dynamic route x 2,836,027 ops/sec ±2.04% (84 runs sampled)
find dynamic multi-parametric route x 1,544,277 ops/sec ±0.50% (93 runs sampled)
find dynamic multi-parametric route with regex x 1,254,514 ops/sec ±1.55% (92 runs sampled)
find long static route x 3,342,545 ops/sec ±1.52% (91 runs sampled)
find long dynamic route x 2,228,427 ops/sec ±1.57% (87 runs sampled)
find static versioned route x 7,022,585 ops/sec ±0.76% (92 runs sampled)
```
 
after:
```
> find-my-way@1.17.0 bench /Users/pp/Projects/PavelPolyakov/find-my-way
> node bench.js
 
lookup static route x 29,220,825 ops/sec ±0.68% (86 runs sampled)
lookup dynamic route x 2,263,062 ops/sec ±0.65% (95 runs sampled)
lookup dynamic multi-parametric route x 1,215,212 ops/sec ±0.95% (93 runs sampled)
lookup dynamic multi-parametric route with regex x 1,102,048 ops/sec ±0.48% (93 runs sampled)
lookup long static route x 1,971,841 ops/sec ±1.19% (93 runs sampled)
lookup long dynamic route x 1,221,743 ops/sec ±2.39% (79 runs sampled)
lookup static versioned route x 5,557,516 ops/sec ±1.80% (89 runs sampled)
find static route x 31,521,768 ops/sec ±1.97% (88 runs sampled)
find dynamic route x 3,173,258 ops/sec ±0.83% (86 runs sampled)
find dynamic multi-parametric route x 1,604,840 ops/sec ±0.78% (89 runs sampled)
find dynamic multi-parametric route with regex x 1,323,553 ops/sec ±0.67% (94 runs sampled)
find long static route x 3,583,129 ops/sec ±0.63% (88 runs sampled)
find long dynamic route x 2,458,582 ops/sec ±0.81% (91 runs sampled)
find static versioned route x 7,257,725 ops/sec ±0.65% (92 runs sampled)
```

I'll update the PR according to your comments and, it seems you are fine with the approach - then I update readmes as well.

What about dropping `node 4` in favour of "nicer" functions through the destructuring? (last question from the [first comment](https://github.com/delvedor/find-my-way/pull/107#issue-240894871))