@mcollina 2019-01-02
Can you rerun the benchmarks once my comments have been addressed? Thanks.

> What about dropping node 4 in favour of "nicer" functions through the destructuring?

Not in this PR. I think @delvedor would be very happy of such a PR and drop Node 4 and 6 in the next major version as long as it does not degrade the benchmarks.