@PavelPolyakov 2019-01-02
before:
```
node bench.js                                                                                                                  [master]
lookup static route x 28,007,182 ops/sec ±1.20% (90 runs sampled)
lookup dynamic route x 2,059,862 ops/sec ±1.39% (88 runs sampled)
lookup dynamic multi-parametric route x 1,133,621 ops/sec ±0.89% (88 runs sampled)
lookup dynamic multi-parametric route with regex x 1,029,175 ops/sec ±0.71% (86 runs sampled)
lookup long static route x 1,894,081 ops/sec ±0.67% (90 runs sampled)
lookup long dynamic route x 1,350,084 ops/sec ±1.00% (89 runs sampled)
lookup static versioned route x 5,489,845 ops/sec ±0.86% (88 runs sampled)
find static route x 30,341,638 ops/sec ±0.69% (91 runs sampled)
find dynamic route x 2,902,873 ops/sec ±0.72% (91 runs sampled)
find dynamic multi-parametric route x 1,452,221 ops/sec ±0.99% (89 runs sampled)
find dynamic multi-parametric route with regex x 1,232,072 ops/sec ±0.65% (88 runs sampled)
find long static route x 3,210,110 ops/sec ±1.06% (92 runs sampled)
find long dynamic route x 2,196,696 ops/sec ±0.73% (88 runs sampled)
find static versioned route x 6,736,033 ops/sec ±0.82% (93 runs sampled)
```

after:
```
node bench.js                                                                                          [feature/versioning-as-strategy]
lookup static route x 28,008,889 ops/sec ±1.20% (89 runs sampled)
lookup dynamic route x 2,088,701 ops/sec ±1.28% (90 runs sampled)
lookup dynamic multi-parametric route x 1,145,452 ops/sec ±0.92% (90 runs sampled)
lookup dynamic multi-parametric route with regex x 788,203 ops/sec ±10.53% (71 runs sampled)
lookup long static route x 1,891,717 ops/sec ±0.86% (91 runs sampled)
lookup long dynamic route x 1,302,576 ops/sec ±1.03% (88 runs sampled)
lookup static versioned route x 5,400,526 ops/sec ±0.73% (90 runs sampled)
find static route x 29,781,526 ops/sec ±1.72% (89 runs sampled)
find dynamic route x 2,806,925 ops/sec ±0.86% (91 runs sampled)
find dynamic multi-parametric route x 1,436,457 ops/sec ±0.79% (93 runs sampled)
find dynamic multi-parametric route with regex x 1,200,777 ops/sec ±1.20% (91 runs sampled)
find long static route x 3,227,337 ops/sec ±0.44% (94 runs sampled)
find long dynamic route x 2,198,234 ops/sec ±0.33% (93 runs sampled)
find static versioned route x 6,355,204 ops/sec ±0.78% (94 runs sampled)
```