@PavelPolyakov 2019-01-02
done and squashed everything, could you, please, review them briefly.
Especially this one: https://github.com/delvedor/find-my-way/pull/107/files#diff-b52768974e6bc0faccb7d4b75b162c99R78 .