@PavelPolyakov 2019-01-02
Thanks for your corrections, I hope you don't mind I've squashed them into one commit.