@PavelPolyakov 2019-01-07
@delvedor , @mcollina 
I've pushed minor new changes, that one can declare not only `version` but `versions` as well.

Here are the benchmarks (with node v10):

before:
```
lookup static route x 47,166,959 ops/sec ±0.71% (92 runs sampled)
lookup dynamic route x 3,367,724 ops/sec ±1.66% (93 runs sampled)
lookup dynamic multi-parametric route x 1,805,423 ops/sec ±1.52% (90 runs sampled)
lookup dynamic multi-parametric route with regex x 1,406,466 ops/sec ±1.19% (95 runs sampled)
lookup long static route x 3,128,009 ops/sec ±2.19% (90 runs sampled)
lookup long dynamic route x 2,209,199 ops/sec ±2.18% (90 runs sampled)
lookup static versioned route x 8,312,062 ops/sec ±1.62% (92 runs sampled)
find static route x 39,353,496 ops/sec ±0.58% (89 runs sampled)
find dynamic route x 4,378,003 ops/sec ±0.75% (91 runs sampled)
find dynamic multi-parametric route x 2,213,928 ops/sec ±1.18% (92 runs sampled)
find dynamic multi-parametric route with regex x 1,545,001 ops/sec ±1.01% (92 runs sampled)
find long static route x 4,527,562 ops/sec ±0.87% (92 runs sampled)
find long dynamic route x 3,202,968 ops/sec ±0.97% (92 runs sampled)
find static versioned route x 10,292,764 ops/sec ±4.08% (86 runs sampled)
```

after:
```
lookup static route x 45,906,750 ops/sec ±2.03% (88 runs sampled)
lookup dynamic route x 3,370,274 ops/sec ±1.81% (90 runs sampled)
lookup dynamic multi-parametric route x 1,815,751 ops/sec ±2.10% (88 runs sampled)
lookup dynamic multi-parametric route with regex x 1,397,260 ops/sec ±1.47% (90 runs sampled)
lookup long static route x 3,224,576 ops/sec ±0.63% (91 runs sampled)
lookup long dynamic route x 2,276,607 ops/sec ±1.00% (94 runs sampled)
lookup static versioned route x 8,500,622 ops/sec ±0.70% (93 runs sampled)
find static route x 39,295,821 ops/sec ±0.89% (90 runs sampled)
find dynamic route x 4,162,268 ops/sec ±1.13% (90 runs sampled)
find dynamic multi-parametric route x 2,166,140 ops/sec ±1.51% (92 runs sampled)
find dynamic multi-parametric route with regex x 1,592,641 ops/sec ±1.04% (92 runs sampled)
find long static route x 4,795,387 ops/sec ±1.05% (91 runs sampled)
find long dynamic route x 3,179,577 ops/sec ±1.07% (90 runs sampled)
find static versioned route x 10,933,767 ops/sec ±1.30% (91 runs sampled)
```

Let me know if you think this should not be merged together, I can rollback.

@delvedor it would be great if you can review this MR, so I can start integrating these changes into `fastify`.